+++ spec/spec_helper.rb
@@ -1,7 +1,8 @@
 require 'rspec/support/spec'
+require 'rspec/mocks/ruby_features'
 
 RSpec::Support::Spec.setup_simplecov do
-  minimum_coverage 97
+  minimum_coverage 96
 end
 
 require 'yaml'
@@ -76,6 +77,8 @@ RSpec.configure do |config|
 
   config.include VerifyAndResetHelpers
   config.include VerificationHelpers
+  config.extend RSpec::Mocks::RubyFeatures
+  config.include RSpec::Mocks::RubyFeatures
 end
 
 shared_context "with syntax" do |syntax|
@@ -115,5 +118,4 @@ shared_context "with the default mocks syntax" do
   after(:all) do
     RSpec::Mocks.configuration.syntax = orig_syntax
   end
-
 end